Drop unneeded cast in NudgePatch constructor

+++ cinelerra/patchgui.C
@@ -571,7 +571,7 @@ NudgePatch::NudgePatch(MWindow *mwindow,
 	y,
 	w,
 	1,
-	(float)patch->track->nudge)
+	patch->track->nudge)
 {
 	this->mwindow = mwindow;
 	this->patch = patch;
